Move common DCAP types to dedicated crate
The `mc-sgx-dcap-sys-types` crate as been created to house common dcap
types used between all the DCAP FFI crates.
