+++ dcap/sys/types/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "mc-sgx-dcap-quoteverify-sys-types"
+name = "mc-sgx-dcap-sys-types"
 version = "0.1.0"
 authors = ["MobileCoin"]
 categories = ["external-ffi-bindings", "hardware-support", "no-std"]
-description = "FFI type definitions used by the `sgx_dcap_quoteverify` library."
+description = "FFI type definitions for the SGX DCAP libraries."
 edition = "2021"
 keywords = ["ffi", "sgx", "no-std"]
 license = "Apache-2.0"
@@ -12,8 +12,8 @@ repository = "https://github.com/mobilecoinfoundation/sgx"
 rust-version = "1.62.1"
 
 [dependencies]
-mc-sgx-dcap-sys-types = { path = "../../../dcap/sys/types", version = "=0.1.0" }
+mc-sgx-core-sys-types = { path = "../../../core/sys/types", version = "=0.1.0" }
 
 [build-dependencies]
 bindgen = "0.60.1"
-mc-sgx-core-build = { path = "../../../core/build" }
+mc-sgx-core-build = { path = "../../../core/build", version = "=0.1.0" }
