+++ dcap/sys/types/wrapper.h
@@ -1,0 +1,5 @@
+#include <sgx_ql_quote.h>
+#include <sgx_ql_lib_common.h>
+#include <sgx_pce.h>
+#include <sgx_quote_3.h>
+#include <sgx_qve_header.h>
