+++ dcap_ql/sys/Cargo.toml
@@ -13,10 +13,11 @@ repository = "https://github.com/mobilecoinfoundation/sgx"
 rust-version = "1.62.1"
 
 [dependencies]
-mc-sgx-dcap-ql-sys-types = { path = "types", version = "=0.1.0" }
 mc-sgx-core-sys-types = { path = "../../core/sys/types", version = "=0.1.0" }
+mc-sgx-dcap-sys-types = { path = "../../dcap/sys/types", version = "=0.1.0" }
+mc-sgx-dcap-ql-sys-types = { path = "types", version = "=0.1.0" }
 
 [build-dependencies]
 bindgen = "0.60.1"
 cargo-emit = "0.2.1"
-mc-sgx-core-build = { path = "../../core/build" }
+mc-sgx-core-build = { path = "../../core/build", version = "=0.1.0" }
