+++ dcap_ql/sys/types/Cargo.toml
@@ -11,9 +11,6 @@ readme = "README.md"
 repository = "https://github.com/mobilecoinfoundation/sgx"
 rust-version = "1.62.1"
 
-[dependencies]
-mc-sgx-core-sys-types = { path = "../../../core/sys/types", version = "=0.1.0" }
-
 [build-dependencies]
 bindgen = "0.60.1"
-mc-sgx-core-build = { path = "../../../core/build" }
+mc-sgx-core-build = { path = "../../../core/build", version = "=0.1.0" }
