+++ dcap_ql/sys/types/wrapper.h
@@ -1,3 +1,1 @@
 #include <sgx_dcap_ql_wrapper.h>
-#include <sgx_ql_lib_common.h>
-#include <sgx_pce.h>
