+++ dcap_quoteverify/sys/types/wrapper.h
@@ -1,4 +1,1 @@
 #include <sgx_dcap_quoteverify.h>
-#include <sgx_qve_header.h>
-#include <sgx_ql_quote.h>
-#include <sgx_quote_3.h>
